add ts testis
