+++ packages/ui/src/components/vuestic-components/va-checkbox/VaCheckbox.spec.js
@@ -3,7 +3,7 @@ import { shallowMount } from '@vue/test-utils'
 import VaCheckbox from './VaCheckbox'
 
 import { ColorThemePlugin } from '../../../services/ColorThemePlugin'
-import { testIsSelectableComponent } from '../../vuestic-mixins/testIsSelectableComponent'
+import { testIsSelectableComponent } from '../../vuestic-mixins/SelectableMixin/testIsSelectableComponentectableComponent'
 
 Vue.use(ColorThemePlugin)
