+++ packages/ui/src/components/vuestic-components/va-switch/VaSwitch.vue
@@ -71,7 +71,7 @@
 </template>
 <script>
 import VaProgressCircle from '../va-progress-bar/progress-types/VaProgressCircle'
-import { SelectableMixin } from '../../vuestic-mixins/SelectableComponent/SelectableMixin'
+import { SelectableMixin } from '../../vuestic-mixins/SelectableMixin/SelectableMixin'
 import { makeContextablePropsMixin } from '../../context-test/context-provide/ContextPlugin'
 import { getColor } from '../../../services/ColorThemePlugin'
 import VaInputWrapper from '../va-input/VaInputWrapper'
